Removed Empty lines as per instructions, moved dto out of /common to /data.
