+++ lib/build.gradle
@@ -28,12 +28,9 @@ dependencies {
     implementation "io.ktor:ktor-server-netty:$ktor_version"
     implementation 'org.junit.jupiter:junit-jupiter:5.4.2'
     implementation "org.jetbrains.kotlinx:kotlinx-serialization-json:1.0.0-RC2"
-
     //GSON
     implementation 'com.google.code.gson:gson:2.8.6'
     testImplementation "io.mockk:mockk:$mockk_version"
-
-
 }
 
 sourceCompatibility = "1.8"
