+++ lib/src/main/java/com/thenewboston/data/dto/bankapi/accountdto/AccountDTO.kt
@@ -1,15 +1,15 @@
-package com.thenewboston.common.http.bankapi.validatorconfirmationservicesdto
+package com.thenewboston.data.dto.bankapi.accountdto
 
 import com.google.gson.annotations.SerializedName
 import java.util.*
 
-data class ValidatorConfirmationServicesDTO(
+data class AccountDTO(
     val id: String,
     @SerializedName("created_date")
     val createdDate: Date,
     @SerializedName("modified_date")
     val modifiedDate: Date,
-    val end: Date,
-    val start: Date,
-    val validator: String
+    @SerializedName("account_number")
+    val accountNumber: String,
+    val trust: Double
 )
