+++ lib/src/main/java/com/thenewboston/data/dto/bankapi/banktransactiondto/BankTransactionDTO.kt
@@ -1,12 +1,10 @@
-package com.thenewboston.common.http.bankapi.banktransactiondto
-
+package com.thenewboston.data.dto.bankapi.banktransactiondto
 
 data class BankTransactionDTO(
     val id: String,
     val block: BlockDTO,
     val amount: Double,
     val recipient: String
-
 )
