+++ lib/src/main/java/com/thenewboston/data/dto/bankapi/banktransactiondto/BlockDTO.kt
@@ -1,10 +1,9 @@
-package com.thenewboston.common.http.bankapi.banktransactiondto
+package com.thenewboston.data.dto.bankapi.banktransactiondto
 
 import com.google.gson.annotations.SerializedName
 import java.util.*
 
 data class BlockDTO(
-
     val id: String,
     @SerializedName("created_date")
     val createdDate: Date,
@@ -14,6 +13,4 @@ data class BlockDTO(
     val balanceKey: String,
     val sender: String,
     val signature: String
-
-
 )
