+++ lib/src/main/java/com/thenewboston/data/dto/bankapi/configdto/ConfigDTO.kt
@@ -1,8 +1,26 @@
-package com.thenewboston.common.http.bankapi.validatordto
+package com.thenewboston.data.dto.bankapi.configdto
 
 import com.google.gson.annotations.SerializedName
 
-data class ValidatorDTO(
+data class ConfigDTO(
+    @SerializedName("primary_validator")
+    val primaryValidator: PrimaryValidator,
+    @SerializedName("account_number")
+    val accountNumber: String,
+    @SerializedName("ip_address")
+    val ipAddress: String,
+    @SerializedName("node_identifier")
+    val nodeIdentifier: String,
+    val port: Int,
+    val protocol: String,
+    val version: String,
+    @SerializedName("default_transaction_fee")
+    val defaultTransactionFee: Double,
+    @SerializedName("node_type")
+    val nodeType: String
+)
+
+data class PrimaryValidator(
     @SerializedName("account_number")
     val accountNumber: String,
     @SerializedName("ip_address")
@@ -23,5 +41,4 @@ data class ValidatorDTO(
     @SerializedName("daily_confirmation_rate")
     val dailyConfirmationRate: Double? = null,
     val trust: Double
-
 )
