+++ lib/src/main/java/com/thenewboston/data/dto/bankapi/invalidblockdto/InvalidBlockDTOTest.kt
@@ -1,12 +1,10 @@
-package com.thenewboston.common.http.bankapi.invalidblockdto
+package com.thenewboston.data.dto.bankapi.invalidblockdto
 
 import com.google.gson.Gson
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 
 class InvalidBlockDTOTest {
-
-
     private val jsonString = """
 {
     "id": "2bcd53c5-19f9-4226-ab04-3dfb17c3a1fe",
@@ -21,7 +19,6 @@ class InvalidBlockDTOTest {
 
     @Test
     fun invalidBlockTest() {
-
         val invalidBlock = Gson().fromJson(jsonString, InvalidBlockDTO::class.java)
 
         invalidBlock.apply {
@@ -35,5 +32,4 @@ class InvalidBlockDTOTest {
             assertEquals("51461a75-dd8d-4133-81f4-543a3b054149", primaryValidator)
         }
     }
-
 }
