+++ lib/src/main/java/com/thenewboston/data/dto/bankapi/validatorconfirmationservicesdto/ValidatorConfirmationServicesDTO.kt
@@ -1,17 +1,15 @@
-package com.thenewboston.common.http.bankapi.accountdto
+package com.thenewboston.data.dto.bankapi.validatorconfirmationservicesdto
 
 import com.google.gson.annotations.SerializedName
 import java.util.*
 
-data class AccountDTO(
-
+data class ValidatorConfirmationServicesDTO(
     val id: String,
     @SerializedName("created_date")
     val createdDate: Date,
     @SerializedName("modified_date")
     val modifiedDate: Date,
-    @SerializedName("account_number")
-    val accountNumber: String,
-    val trust: Double
-
+    val end: Date,
+    val start: Date,
+    val validator: String
 )
