+++ lib/src/main/java/com/thenewboston/data/dto/bankapi/validatordto/ValidatorDTO.kt
@@ -1,28 +1,8 @@
-package com.thenewboston.common.http.bankapi.configdto
+package com.thenewboston.data.dto.bankapi.validatordto
 
 import com.google.gson.annotations.SerializedName
 
-data class ConfigDTO(
-    @SerializedName("primary_validator")
-    val primaryValidator: PrimaryValidator,
-    @SerializedName("account_number")
-    val accountNumber: String,
-    @SerializedName("ip_address")
-    val ipAddress: String,
-    @SerializedName("node_identifier")
-    val nodeIdentifier: String,
-    val port: Int,
-    val protocol: String,
-    val version: String,
-    @SerializedName("default_transaction_fee")
-    val defaultTransactionFee: Double,
-    @SerializedName("node_type")
-    val nodeType: String
-
-)
-
-data class PrimaryValidator(
-
+data class ValidatorDTO(
     @SerializedName("account_number")
     val accountNumber: String,
     @SerializedName("ip_address")
@@ -43,6 +23,4 @@ data class PrimaryValidator(
     @SerializedName("daily_confirmation_rate")
     val dailyConfirmationRate: Double? = null,
     val trust: Double
-
-
 )
